+++ covenant-signer/cmd/keys.go
@@ -1,0 +1,9 @@
+package cmd
+
+import (
+	"github.com/cosmos/cosmos-sdk/client/keys"
+)
+
+func init() {
+	rootCmd.AddCommand(keys.Commands())
+}